pyproject.toml
[tool.poetry]
name = "fast-slides"
version = "0.1.0"
description = "A fast and elegant slide generator from Markdown"
authors = ["zweix123 <1979803044@qq.com>"]
readme = "README.md"
package-mode = true

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.2"
jinja2 = "^3.1.2"
markdown = "^3.4.1"
pyquery = "^2.0.0"
pygments = "^2.14.0"
lxml = "^5.0.0"
pyyaml = "^6.0.1"
click = "^8.1.3"
watchdog = "^3.0.0"

[tool.poetry.group.dev.dependencies]
types-markdown = "^3.4.2.6"
types-requests = "^2.28.11.17"
types-pyyaml = "^6.0.12.12"
types-click = "^8.1.7.13"
black = "^23.1.0"
mypy = "^1.1.1"

[tool.poetry.scripts]
fast-slides = "fast_slides.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
